Add files via upload
School Database is a java designed application which helps to manage and analyze school system. It  has a menu which interacts with a user, and helps to add new students, faculty, employee and to  add or modify information about them( such info as Name, name of the course, number of credits and etc.) The program analyzes the data and sort it accordingly as well.